requirements.txt
os
json
redis
pandas
pathlib
pymongo
MySQLdb
datetime
requests
sqlalchemy
python-dotenv